+++ stingray/powerspectrum.py
@@ -8,8 +8,8 @@ import scipy.optimize
 import stingray.lightcurve as lightcurve
 import stingray.utils as utils
 from stingray.gti import bin_intervals_from_gtis, check_gtis
-from stingray.utils import simon
 from stingray.crossspectrum import Crossspectrum, AveragedCrossspectrum
+from stingray.stats import pds_probability
 
 try:
     from tqdm import tqdm as show_progress
@@ -18,122 +18,6 @@ except ImportError:
         return a
 
 __all__ = ["Powerspectrum", "AveragedPowerspectrum", "DynamicalPowerspectrum"]
-
-
-def classical_pvalue(power, nspec):
-    """
-    Compute the probability of detecting the current power under
-    the assumption that there is no periodic oscillation in the data.
-
-    This computes the single-trial p-value that the power was
-    observed under the null hypothesis that there is no signal in
-    the data.
-
-    Important: the underlying assumptions that make this calculation valid
-    are:
-
-    1. the powers in the power spectrum follow a chi-square distribution
-    2. the power spectrum is normalized according to [Leahy 1983]_, such
-       that the powers have a mean of 2 and a variance of 4
-    3. there is only white noise in the light curve. That is, there is no
-       aperiodic variability that would change the overall shape of the power
-       spectrum.
-
-    Also note that the p-value is for a *single trial*, i.e. the power
-    currently being tested. If more than one power or more than one power
-    spectrum are being tested, the resulting p-value must be corrected for the
-    number of trials (Bonferroni correction).
-
-    Mathematical formulation in [Groth 1975]_.
-    Original implementation in IDL by Anna L. Watts.
-
-    Parameters
-    ----------
-    power :  float
-        The squared Fourier amplitude of a spectrum to be evaluated
-
-    nspec : int
-        The number of spectra or frequency bins averaged in ``power``.
-        This matters because averaging spectra or frequency bins increases
-        the signal-to-noise ratio, i.e. makes the statistical distributions
-        of the noise narrower, such that a smaller power might be very
-        significant in averaged spectra even though it would not be in a single
-        power spectrum.
-
-    Returns
-    -------
-    pval : float
-        The classical p-value of the observed power being consistent with
-        the null hypothesis of white noise
-
-    References
-    ----------
-
-    * .. [Leahy 1983] https://ui.adsabs.harvard.edu/#abs/1983ApJ...266..160L/abstract
-    * .. [Groth 1975] https://ui.adsabs.harvard.edu/#abs/1975ApJS...29..285G/abstract
-
-    """
-    if not np.isfinite(power):
-        raise ValueError("power must be a finite floating point number!")
-
-    if power < 0:
-        raise ValueError("power must be a positive real number!")
-
-    if not np.isfinite(nspec):
-        raise ValueError("nspec must be a finite integer number")
-
-    if nspec < 1:
-        raise ValueError("nspec must be larger or equal to 1")
-
-    if not np.isclose(nspec % 1, 0):
-        raise ValueError("nspec must be an integer number!")
-
-    # If the power is really big, it's safe to say it's significant,
-    # and the p-value will be nearly zero
-    if (power * nspec) > 30000:
-        simon("Probability of no signal too miniscule to calculate.")
-        return 0.0
-
-    else:
-        pval = _pavnosigfun(power, nspec)
-        return pval
-
-
-def _pavnosigfun(power, nspec):
-    """
-    Helper function doing the actual calculation of the p-value.
-
-    Parameters
-    ----------
-    power : float
-        The measured candidate power
-
-    nspec : int
-        The number of power spectral bins that were averaged in `power`
-        (note: can be either through averaging spectra or neighbouring bins)
-    """
-    sum = 0.0
-    m = nspec - 1
-
-    pn = power * nspec
-
-    while m >= 0:
-
-        s = 0.0
-        for i in range(int(m) - 1):
-            s += np.log(float(m - i))
-
-        logterm = m * np.log(pn / 2) - pn / 2 - s
-        term = np.exp(logterm)
-        ratio = sum / term
-
-        if ratio > 1.0e15:
-            return sum
-
-        sum += term
-        m -= 1
-
-    return sum
 
 
 class Powerspectrum(Crossspectrum):
@@ -362,19 +246,20 @@ class Powerspectrum(Crossspectrum):
             raise ValueError("This method only works on "
                              "Leahy-normalized power spectra!")
 
+        if trial_correction:
+            ntrial = self.power.shape[0]
+        else:
+            ntrial = 1
+
         if np.size(self.m) == 1:
             # calculate p-values for all powers
             # leave out zeroth power since it just encodes the number of photons!
-            pv = np.array([classical_pvalue(power, self.m)
-                           for power in self.power])
+            pv = pds_probability(self.power, n_summed_spectra=self.m,
+                                 ntrial=ntrial)
         else:
-            pv = np.array([classical_pvalue(power, m)
+            pv = np.array([pds_probability(power, n_summed_spectra=m,
+                                           ntrial=ntrial)
                            for power, m in zip(self.power, self.m)])
-
-        # if trial correction is used, then correct the threshold for
-        # the number of powers in the power spectrum
-        if trial_correction:
-            threshold /= self.power.shape[0]
 
         # need to add 1 to the indices to make up for the fact that
         # we left out the first power above!
