+++ stingray/tests/test_stats.py
@@ -55,3 +55,116 @@ def test_zn_probability(ntrial):
     detlev = z2_n_detection_level(2, 0.1, ntrial=ntrial)
     np.testing.assert_almost_equal(z2_n_probability(detlev, 2, ntrial=ntrial),
                                    0.1)
+
+
+class TestClassicalSignificances(object):
+    def test_function_runs(self):
+        power = 2.0
+        nspec = 1.0
+        with pytest.warns(DeprecationWarning):
+            classical_pvalue(power, nspec)
+
+    def test_power_is_not_infinite(self):
+        power = np.inf
+        nspec = 1
+        with pytest.warns(DeprecationWarning):
+            with pytest.raises(ValueError):
+                classical_pvalue(power, nspec)
+
+    def test_power_is_not_infinite2(self):
+        power = -np.inf
+        nspec = 1
+        with pytest.warns(DeprecationWarning):
+            with pytest.raises(ValueError):
+                classical_pvalue(power, nspec)
+
+    def test_power_is_non_nan(self):
+        power = np.nan
+        nspec = 1
+        with pytest.warns(DeprecationWarning):
+            with pytest.raises(ValueError):
+                classical_pvalue(power, nspec)
+
+    def test_power_is_positive(self):
+        power = -2.0
+        nspec = 1.0
+        with pytest.warns(DeprecationWarning):
+            with pytest.raises(ValueError):
+                classical_pvalue(power, nspec)
+
+    def test_nspec_is_not_infinite(self):
+        power = 2.0
+        nspec = np.inf
+        with pytest.warns(DeprecationWarning):
+            with pytest.raises(ValueError):
+                classical_pvalue(power, nspec)
+
+    def test_nspec_is_not_infinite2(self):
+        power = 2.0
+        nspec = -np.inf
+        with pytest.raises(ValueError):
+            classical_pvalue(power, nspec)
+
+    def test_nspec_is_not_nan(self):
+        power = 2.0
+        nspec = np.nan
+        with pytest.warns(DeprecationWarning):
+            with pytest.raises(ValueError):
+                classical_pvalue(power, nspec)
+
+    def test_nspec_is_positive(self):
+        power = 2.0
+        nspec = -1.0
+        with pytest.warns(DeprecationWarning):
+            with pytest.raises(ValueError):
+                classical_pvalue(power, nspec)
+
+    def test_nspec_is_nonzero(self):
+        power = 2.0
+        nspec = 0.0
+        with pytest.warns(DeprecationWarning):
+            with pytest.raises(ValueError):
+                classical_pvalue(power, nspec)
+
+    def test_nspec_is_an_integer_number(self):
+        power = 2.0
+        nspec = 2.5
+        with pytest.warns(DeprecationWarning):
+            with pytest.raises(ValueError):
+                classical_pvalue(power, nspec)
+
+    def test_nspec_float_type_okay(self):
+        power = 2.0
+        nspec = 2.0
+        with pytest.warns(DeprecationWarning):
+            classical_pvalue(power, nspec)
+
+    def test_pvalue_decreases_with_increasing_power(self):
+        power1 = 2.0
+        power2 = 20.0
+        nspec = 1.0
+        with pytest.warns(DeprecationWarning):
+            pval1 = classical_pvalue(power1, nspec)
+        with pytest.warns(DeprecationWarning):
+            pval2 = classical_pvalue(power2, nspec)
+
+        assert pval1 - pval2 > 0.0
+
+    def test_pvalue_must_decrease_with_increasing_nspec(self):
+        power = 3.0
+        nspec1 = 1.0
+        nspec2 = 10.0
+
+        with pytest.warns(DeprecationWarning):
+            pval1 = classical_pvalue(power, nspec1)
+        with pytest.warns(DeprecationWarning):
+            pval2 = classical_pvalue(power, nspec2)
+
+        assert pval1 - pval2 > 0.0
+
+    def test_very_large_powers_produce_zero_prob(self):
+        power = 31000.0
+        nspec = 1
+        with pytest.warns(DeprecationWarning):
+            pval = classical_pvalue(power, nspec)
+        assert np.isclose(pval, 0.0)
